backend/mainAgent/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.10.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
Deprecated==1.3.1
distro==1.9.0
Django==5.2.7
djangorestframework==3.16.1
filelock==3.20.0
fsspec==2025.10.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
Jinja2==3.1.6
jiter==0.11.1
joblib==1.5.2
lxml==6.0.2
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.5
numpy==2.3.4
openai==2.6.1
packaging==25.0
pdf2image==1.17.0
pdfminer.six==20250506
pdfplumber==0.11.7
pikepdf==10.0.0
pillow==12.0.0
pycparser==2.23
pydantic==2.12.3
pydantic_core==2.41.4
PyPDF2==3.0.1
pypdfium2==5.0.0
pytesseract==0.3.13
python-dotenv==1.2.1
PyYAML==6.0.3
regex==2025.10.23
requests==2.32.5
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.3
sentence-transformers==5.1.2
setuptools==80.9.0
sniffio==1.3.1
sqlparse==0.5.3
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.1
torch==2.9.0
tqdm==4.67.1
transformers==4.57.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
wheel==0.45.1
wrapt==2.0.0
google-generativeai>=0.3.0
